server/requirements.txt
click==7.1.2
Faker==6.5.0
fastapi==0.63.0
h11==0.12.0
mysql-connector-python==8.0.23
pkg-resources==0.0.0
protobuf==3.15.2
pydantic==1.7.3
python-dateutil==2.8.1
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
text-unidecode==1.3
typing-extensions==3.7.4.3
uvicorn==0.13.4
